pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-mem"
version = "0.4.0"
description = "AI memory as a typed knowledge graph in Obsidian. Like Mem0, but you own your data."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ali Baizhanov"},
]
keywords = [
    "memory", "obsidian", "knowledge-graph", "llm", "ai", "mcp",
    "second-brain", "rag", "semantic-search", "embeddings",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.0"]
embeddings = ["sentence-transformers>=2.2"]
mcp = ["mcp>=1.0"]
all = [
    "anthropic>=0.40",
    "openai>=1.0",
    "sentence-transformers>=2.2",
    "mcp>=1.0",
]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.urls]
Homepage = "https://github.com/alibaizhanov/obsidian-mem"
Repository = "https://github.com/alibaizhanov/obsidian-mem"
Issues = "https://github.com/alibaizhanov/obsidian-mem/issues"
Documentation = "https://github.com/alibaizhanov/obsidian-mem#readme"

[project.scripts]
obsidian-mem = "cli:main"

[tool.setuptools.packages.find]
include = ["engine*", "api*"]

[tool.setuptools]
py-modules = ["obsidian_mem", "obsidian_mem_middleware", "cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
